Fixed portability issue within Launch.java
This fixes an incompatibility issue within the Launch class. The issue is that although URLs are separated with semicolons on Microsoft Windows, colons are used on UNIX-based system (macOS and Linux for instance). This causes such systems to not be able to launch the LaunchWrapper properly. This issue is resolved by using File.pathSeparator instead of ";" within line 34.